rustls 0.23

+++ src/io.rs
@@ -376,7 +376,7 @@ where
 
                                     let mut inner = this.inner.borrow_mut();
                                     let response_idx =
-                                        inner.base.wrapping_add(inner.queue.len() as usize);
+                                        inner.base.wrapping_add(inner.queue.len());
 
                                     if let Poll::Ready(res) = res {
                                         // check if current result is only response atm
@@ -398,7 +398,7 @@ where
                                 } else {
                                     let mut inner = this.inner.borrow_mut();
                                     let response_idx =
-                                        inner.base.wrapping_add(inner.queue.len() as usize);
+                                        inner.base.wrapping_add(inner.queue.len());
                                     inner.queue.push_back(ServiceResult::Pending);
 
                                     let st = this.state.clone();
